+++ tools/catalog-local.js
@@ -15,10 +15,13 @@ var packageCache = require('./package-cache.js');
 var PackageSource = require('./package-source.js');
 var VersionParser = require('./package-version-parser.js');
 
-// LocalCatalog represents the packages located into an application folder
-// A default instance of this catalog is created in catalog.js
+// LocalCatalog represents packages located in the application's
+// package directory, other package directories specified via an
+// environment variable, and core packages in the repo if meteor is
+// being run from a git checkout.
 var LocalCatalog = function (options) {
   var self = this;
+  options = options || {};
 
   // Package server data.  Maps package name to a {packageSource, packageRecord,
   // versionRecord, buildRecord} object.
@@ -27,7 +30,7 @@ var LocalCatalog = function (options) {
   // We use the initialization design pattern because it makes it easier to use
   // both of our catalogs as singletons.
   self.initialized = false;
-  self.containingCatalog = options ? options.containingCatalog : self;
+  self.containingCatalog = options.containingCatalog || self;
 
     // Local directories to search for package source trees
   self.localPackageSearchDirs = null;
@@ -44,8 +47,12 @@ var LocalCatalog = function (options) {
   // need to process.
   self.effectiveLocalPackageDirs = [];
 
-   // Each catalog needs its own package cache.
-  self.packageCache = new packageCache.PackageCache(self.containingCatalog);
+   // If this is a standalone catalog (eg BootstrapCatalogCheckout) it needs its
+   // own PackageCache, but if it's part of the complete catalog it should share
+   // a cache with its container.
+  self.packageCache = options.containingCatalog
+    ? options.containingCatalog.packageCache
+    : new packageCache.PackageCache(self);
 
   self.buildRequested = null;
 
@@ -327,7 +334,6 @@ _.extend(LocalCatalog.prototype, {
             testName: packageSource.testName,
             version: version,
             publishedBy: null,
-            earliestCompatibleVersion: packageSource.earliestCompatibleVersion,
             description: packageSource.metadata.summary,
             dependencies: packageSource.getDependencyMetadata(),
             source: null,
@@ -592,7 +598,8 @@ _.extend(LocalCatalog.prototype, {
         unip = compiler.compile(self.packages[name].packageSource, {
           ignoreProjectDeps: constraintSolverOpts.ignoreProjectDeps
         }).isopack;
-        if (! buildmessage.jobHasMessages()) {
+        if (! buildmessage.jobHasMessages() &&
+            ! self.isopacketBuildingCatalog) {
           // Save the build, for a fast load next time
           try {
             var buildDir = path.join(sourcePath, '.build.'+ name);
@@ -644,6 +651,12 @@ _.extend(LocalCatalog.prototype, {
   _maybeGetUpToDateBuild: function (name, constraintSolverOpts) {
     var self = this;
     buildmessage.assertInCapture();
+
+    // When we build packages for isopackets, we're using different build
+    // settings (eg we're ignoring app packages) so they shouldn't share
+    // the same cache.
+    if (self.isopacketBuildingCatalog)
+      return null;
 
     var sourcePath = self.packages[name].packageSource.sourceRoot;
     var buildDir = path.join(sourcePath, '.build.' + name);