+++ tools/catalog-remote.js
@@ -6,11 +6,10 @@ var auth = require('./auth.js');
 var httpHelpers = require('./http-helpers.js');
 var release = require('./release.js');
 var files = require('./files.js');
-var ServiceConnection = require('./service-connection.js');
 var utils = require('./utils.js');
 var buildmessage = require('./buildmessage.js');
 var compiler = require('./compiler.js');
-var uniload = require('./uniload.js');
+var isopackets = require("./isopackets.js");
 var tropohouse = require('./tropohouse.js');
 var config = require('./config.js');
 var packageClient = require('./package-client.js');
@@ -565,7 +564,8 @@ _.extend(RemoteCatalog.prototype, {
 
   getSortedVersions: function (name) {
     var self = this;
-    var match = this._getPackageVersions(name);
+    var match = this._columnsQuery(
+      "SELECT version FROM versions WHERE packageName=?", name);
     if (match === null)
       return [];
     return _.pluck(match, 'version').sort(VersionParser.compare);
@@ -592,14 +592,6 @@ _.extend(RemoteCatalog.prototype, {
       throw new Error("Found multiple packages matching name: " + name);
     }
     return result[0];
-  },
-
-  _getPackageVersions: function (name) {
-    if (!name) {
-      throw new Error("No name provided");
-    }
-    return this._contentQuery(
-      "SELECT content FROM versions WHERE packageName=?", name);
   },
 
   getAllBuilds: function (name, version) {