+++ History.md
@@ -3,6 +3,17 @@
 * Spacebars: Allow curly braces to be escaped, with special
   sequences `{{|` and `{{{|` to insert a literal `{{` or `{{{`.
 
+* Corporate HTTP proxy support is now implemented using our websocket library's
+  new built-in implementation instead of a custom implementation. #2515
+
+* Some packages are no longer released as part of the core release process:
+  amplify, backbone, bootstrap, d3, jquery-history, and jquery-layout.
+
+* Rework how Meteor packages get loaded into the command-line tool
+
+* Remove support for the undocumented earliestCompatibleVersion feature of the
+  package system.
+
 * Upgraded dependencies:
   - node: 0.10.33 (from 0.10.29)
   - source-map-support: 0.2.8 (from 0.2.5)
@@ -15,6 +26,7 @@
   - http-proxy: 1.6.0 (from a fork of 1.0.2)
   - esprima: 1.2.2 (from an unreleased 1.1-era commit)
   - openssl in mongo: 1.0.1j (from 1.0.1g)
+  - faye-websocket: 0.8.1 (from using websocket-driver instead)
   - MongoDB: 2.4.12 (from 2.4.9)
 
 ## v1.0